pyproject.toml
[tool.poetry]
name = "historia-ai"
version = "0.1.0"
description = ""
authors = ["Jagath Jai Kumar <jagath.jaikumar@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1.3"
django-environ = "^0.11.2"
django-health-check = "^3.18.3"
psycopg2-binary = "^2.9.10"
dj-database-url = "^2.3.0"

# First party apps
historia-indexing = { path = "apps/historia_indexing", develop = true }
djangorestframework = "^3.15.2"

[tool.poetry.group.serve.dependencies]
gunicorn = "*"

[tool.poetry.group.dev.dependencies]
ruff = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
